+++ test/matchRoute.spec.js
@@ -1,7 +1,7 @@
 /**
  * Universal Router (https://www.kriasoft.com/universal-router/)
  *
- * Copyright © 2015-2016 Konstantin Tarkus, Kriasoft LLC. All rights reserved.
+ * Copyright © 2015-present Konstantin Tarkus, Kriasoft LLC. All rights reserved.
  *
  * This source code is licensed under the Apache 2.0 license found in the
  * LICENSE.txt file in the root directory of this source tree.
@@ -10,13 +10,22 @@
 import { expect } from 'chai';
 import matchRoute from '../src/matchRoute';
 
-describe('matchRoute(route, baseUrl, path)', () => {
+function toArray(gen) {
+  const arr = [];
+  let res = gen.next();
+  while (!res.done) {
+    arr.push(res.value);
+    res = gen.next();
+  }
+  return arr;
+}
 
+describe('matchRoute(route, baseUrl, path)', () => {
   it('should match 0 routes (1)', () => {
     const route = {
       path: '/',
     };
-    const result = Array.from(matchRoute(route, '', '/a'));
+    const result = toArray(matchRoute(route, '', '/a'));
     expect(result).to.have.lengthOf(0);
   });
 
@@ -24,7 +33,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
     const route = {
       path: '/a',
     };
-    const result = Array.from(matchRoute(route, '', '/b'));
+    const result = toArray(matchRoute(route, '', '/b'));
     expect(result).to.have.lengthOf(0);
   });
 
@@ -37,7 +46,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
         },
       ],
     };
-    const result = Array.from(matchRoute(route, '', '/b'));
+    const result = toArray(matchRoute(route, '', '/b'));
     expect(result).to.have.lengthOf(0);
   });
 
@@ -45,7 +54,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
     const route = {
       path: '/',
     };
-    const result = Array.from(matchRoute(route, '', '/'));
+    const result = toArray(matchRoute(route, '', '/'));
     expect(result).to.have.lengthOf(1);
     expect(result[0]).to.have.property('baseUrl', '');
     expect(result[0]).to.have.property('path', '/');
@@ -56,7 +65,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
     const route = {
       path: '/a',
     };
-    const result = Array.from(matchRoute(route, '', '/a'));
+    const result = toArray(matchRoute(route, '', '/a'));
     expect(result).to.have.lengthOf(1);
     expect(result[0]).to.have.property('baseUrl', '');
     expect(result[0]).to.have.property('path', '/a');
@@ -72,7 +81,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
         },
       ],
     };
-    const result = Array.from(matchRoute(route, '', '/a'));
+    const result = toArray(matchRoute(route, '', '/a'));
     expect(result).to.have.lengthOf(2);
     expect(result[0]).to.have.property('baseUrl', '');
     expect(result[0]).to.have.property('path', '/');
@@ -96,7 +105,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
         },
       ],
     };
-    const result = Array.from(matchRoute(route, '', '/a/b/c'));
+    const result = toArray(matchRoute(route, '', '/a/b/c'));
     expect(result).to.have.lengthOf(3);
     expect(result[0]).to.have.property('baseUrl', '');
     expect(result[0]).to.have.deep.property('route.path', '/a');
@@ -115,12 +124,11 @@ describe('matchRoute(route, baseUrl, path)', () => {
         },
       ],
     };
-    const result = Array.from(matchRoute(route, '', '/'));
+    const result = toArray(matchRoute(route, '', '/'));
     expect(result).to.have.lengthOf(2);
     expect(result[0]).to.have.property('baseUrl', '');
     expect(result[0]).to.have.deep.property('route.path', '/');
     expect(result[1]).to.have.property('baseUrl', '');
     expect(result[1]).to.have.deep.property('route.path', '/');
   });
-
 });