+++ tools/build.js
@@ -1,13 +1,11 @@
 /**
  * Universal Router (https://www.kriasoft.com/universal-router/)
  *
- * Copyright © 2015-2016 Konstantin Tarkus, Kriasoft LLC. All rights reserved.
+ * Copyright © 2015-present Konstantin Tarkus, Kriasoft LLC. All rights reserved.
  *
  * This source code is licensed under the Apache 2.0 license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
-
-'use strict';
 
 const fs = require('fs');
 const del = require('del');
@@ -19,51 +17,39 @@ const nodeResolve = require('rollup-plugin-node-resolve');
 const pkg = require('../package.json');
 
 // The source files to be compiled by Rollup
-const files = [
+let files = [
   {
     format: 'cjs',
     ext: '.js',
-    presets: ['es2016', 'es2017'],
-    plugins: [
-      'external-helpers',
-      'transform-runtime',
-    ],
+    output: 'main',
   },
   {
     format: 'es',
     ext: '.mjs',
-    presets: ['es2016', 'es2017'],
-    plugins: [
-      'external-helpers',
-      'transform-runtime',
-    ],
+    output: 'main',
   },
   {
     format: 'cjs',
     ext: '.js',
     output: 'legacy',
     presets: [['latest', { es2015: { modules: false } }]],
-    plugins: ['transform-runtime'],
   },
   {
     format: 'cjs',
     ext: '.js',
     output: 'browser',
     presets: [['latest', { es2015: { modules: false } }]],
-    plugins: ['transform-runtime'],
   },
   {
     format: 'es',
     ext: '.mjs',
     output: 'browser',
     presets: [['latest', { es2015: { modules: false } }]],
-    plugins: ['transform-runtime'],
   },
   {
     format: 'umd',
     ext: '.js',
     presets: [['latest', { es2015: { modules: false } }]],
-    plugins: ['transform-runtime'],
     output: pkg.name,
     moduleName: 'UniversalRouter',
   },
@@ -71,12 +57,18 @@ const files = [
     format: 'umd',
     ext: '.min.js',
     presets: [['latest', { es2015: { modules: false } }]],
-    plugins: ['transform-runtime'],
     output: pkg.name,
     moduleName: 'UniversalRouter',
     minify: true,
   },
 ];
+
+// Compile generateUrls module
+files = files.concat(files.map(file => Object.assign({}, file, {
+  entry: 'src/generateUrls.js',
+  output: file.output === pkg.name ? `${pkg.name}-generate-urls` : `generateUrls/${file.output}`,
+  moduleName: file.moduleName && 'generateUrls',
+})));
 
 let promise = Promise.resolve();
 
@@ -86,7 +78,7 @@ promise = promise.then(() => del(['build/*']));
 // Compile source code into a distributable format with Babel
 files.forEach((file) => {
   promise = promise.then(() => rollup.rollup({
-    entry: 'src/main.js',
+    entry: file.entry || 'src/Router.js',
     external: file.format === 'umd' ? [] : Object.keys(pkg.dependencies),
     plugins: [
       ...file.format === 'umd' ? [nodeResolve({ browser: true }), commonjs()] : [],
@@ -100,22 +92,28 @@ files.forEach((file) => {
       ...file.minify ? [uglify()] : [],
     ],
   }).then(bundle => bundle.write({
-    dest: `build/${file.output || 'main'}${file.ext}`,
+    dest: `build/${file.output}${file.ext}`,
     format: file.format,
     sourceMap: !file.minify,
-    exports: 'named',
+    exports: 'default',
     moduleName: file.moduleName,
   })));
 });
 
 // Copy package.json and LICENSE.txt
 promise = promise.then(() => {
-  delete pkg.private;
   delete pkg.devDependencies;
   delete pkg.scripts;
   delete pkg.eslintConfig;
   delete pkg.babel;
+  const pkg2 = Object.assign({}, pkg, {
+    name: 'generateUrls',
+    description: 'Universal Router extension for URLs generation',
+  });
+  delete pkg.private;
+  delete pkg2.dependencies;
   fs.writeFileSync('build/package.json', JSON.stringify(pkg, null, '  '), 'utf-8');
+  fs.writeFileSync('build/generateUrls/package.json', JSON.stringify(pkg2, null, '  '), 'utf-8');
   fs.writeFileSync('build/README.md', fs.readFileSync('README.md', 'utf-8'), 'utf-8');
   fs.writeFileSync('build/LICENSE.txt', fs.readFileSync('LICENSE.txt', 'utf-8'), 'utf-8');
 });